Restructures project
